perf: switch to using a web worker for parsing a save file
re: #176
